# cc-webapp/backend/current_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Casino-Club F2P API",
    "description": "Backend API for Casino-Club F2P gaming platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/rewards/users/{user_id}/rewards": {
      "get": {
        "tags": [
          "Rewards",
          "Rewards",
          "Rewards"
        ],
        "summary": "Get User Rewards",
        "description": "Retrieves a paginated list of rewards for a specific user.",
        "operationId": "get_user_rewards_api_rewards_users__user_id__rewards_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "The ID of the user to get rewards for"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number, 1-indexed",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, 1-indexed"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRewardsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rewards/distribute": {
      "post": {
        "tags": [
          "Rewards",
          "Rewards",
          "rewards"
        ],
        "summary": "Distribute Reward To User",
        "description": "Distributes a specific reward to a user.\nThis is the central endpoint for granting rewards from games or events.",
        "operationId": "distribute_reward_to_user_api_rewards_distribute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RewardDistributionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedRewardsResponse": {
        "properties": {
          "rewards": {
            "items": {
              "$ref": "#/components/schemas/RewardItem"
            },
            "type": "array",
            "title": "Rewards"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_rewards": {
            "type": "integer",
            "title": "Total Rewards"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "rewards",
          "page",
          "page_size",
          "total_rewards",
          "total_pages"
        ],
        "title": "PaginatedRewardsResponse"
      },
...
      "RewardDistributionRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "reward_type": {
            "type": "string",
            "title": "Reward Type"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "source_description": {
            "type": "string",
            "title": "Source Description"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "metadata": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "reward_type",
          "amount",
          "source_description"
        ],
        "title": "RewardDistributionRequest"
      },
      "RewardItem": {
        "properties": {
          "reward_id": {
            "type": "integer",
            "title": "Reward Id"
          },
          "reward_type": {
            "type": "string",
            "title": "Reward Type"
          },
          "reward_value": {
            "type": "string",
            "title": "Reward Value"
          },
          "awarded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Awarded At"
          }
        },
        "type": "object",
        "required": [
          "reward_id",
          "reward_type",
          "reward_value",
          "awarded_at"
        ],
        "title": "RewardItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        }
      }
    }
  }
}